+++ package.json
@@ -8,6 +8,8 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "@tiptap/starter-kit": "^2.0.0-beta.139",
+    "@tiptap/vue-2": "^2.0.0-beta.67",
     "axios": "^0.24.0",
     "core-js": "^3.6.5",
     "node-sass": "^6.0.1",
