+++ src/views/CounselingRoom.vue
@@ -1,32 +1,30 @@
 <template>
-    <!-- 이후 roomId를 조정해 개인실로 만들기 -->
-    <vue-webrtc ref="webrtc" width="100%" roomId="234132"/>
+    <div class="container">
+        <div class="left">
+            <WebRTC />
+        </div>
+        <div class="doctor-noti">
+            
+        </div>
+    </div>
 </template>
 
 <script>
-/* vue-webrtc를 위한 import 및 코드 */
-import Vue from 'vue'
-import WebRTC from 'vue-webrtc'
-import * as io from 'socket.io-client'
-window.io = io
-Vue.use(WebRTC)
-/* vue-webrtc 구동용 코드 끝 */
-
+import WebRTC from '@/components/WebRTC.vue'
 export default {
     name: 'CounselingRoom',
-    mounted: function() {
-        this.$refs.webrtc.join() // vue-webrtc 구동을 위한 함수입니다.
+    components: {
+        WebRTC
     }
 }
 </script>
 
 <style lang="scss" scoped>
-vue-webrtc {
-    font-family: Avenir, Helvetica, Arial, sans-serif;
-    -webkit-font-smoothing: antialiased;
-    -moz-osx-font-smoothing: grayscale;
-    text-align: center;
-    color: #2c3e50;
-    margin-top: 60px;
+.container {
+    display: flex;
+    .left {
+        width: 30%;
+        min-height: 100%;
+    }
 }
 </style>
